fix: refactor history servcie
